backend/requirements.txt
flask
pandas
numpy
scipy
mne
matplotlib
torch
psycopg2-binary
werkzeug
gunicorn